refactor(wallet): move logic to WalletButton
