+++ app/src/lib/WalletButton.svelte
@@ -1,18 +1,41 @@
+<script lang="ts">
+	import { walletStore } from '$utils/walletStore';
+
+	function showAddress(store) {
+		const base58 = store.publicKey?.toBase58();
+		if (!store.wallet || !base58) return null;
+		return base58.slice(0, 4) + '..' + base58.slice(-4);
+	}
+
+	function onButtonClick() {
+		if ($walletStore.connected) {
+			// show list modal
+
+			return;
+		}
+
+		// TODO: show wallet selection modal
+		selectWallet('Phantom');
+	}
+
+	const selectWallet = (walletName) => {
+		$walletStore.select(walletName);
+	};
+
+	$: address = showAddress($walletStore);
+</script>
+
 <button
 	class="wallet-adapter-button"
-	class:justify-between={$$slots['start-icon'] || $$slots['end-icon']}
-	on:click
+	class:justify-between={$walletStore.connected}
+	on:click={() => onButtonClick()}
 >
-	{#if $$slots['start-icon']}
-		<i class="wallet-adapter-button-start-icon">
-			<slot name="start-icon" />
-		</i>
-	{/if}
-	<slot />
-	{#if $$slots['end-icon']}
-		<i class="wallet-adapter-button-end-icon">
-			<slot name="end-icon" />
-		</i>
+	{#if $walletStore.connected}
+		{address}
+
+		<img src={$walletStore.wallet.icon} alt={$walletStore.wallet.name + ' icon'} />
+	{:else}
+		Select wallet
 	{/if}
 </button>
